# Ci20 v3.18 MSC0 Performance Fix