--- a/drivers/clk/jz47xx/jz47xx-cgu.c
+++ b/drivers/clk/jz47xx/jz47xx-cgu.c
@@ -332,6 +332,7 @@ static unsigned long jz47xx_clk_recalc_rate(struct clk_hw *hw,
 		div = (div_reg >> clk_info->div.shift) &
 		      ((1 << clk_info->div.bits) - 1);
 		div += 1;
+		div <<= clk_info->div.div;
 
 		rate /= div;
 	}
@@ -351,6 +352,12 @@ static unsigned jz47xx_clk_calc_div(const struct jz47xx_cgu_clk_info *clk_info,
 	/* and impose hardware constraints */
 	div = min_t(unsigned, div, 1 << clk_info->div.bits);
 	div = max_t(unsigned, div, 1);
+
+	/* If the divider value itself must be divided before being written to
+	 * the divider register, we must ensure we don't have any bits set that
+	 * would be lost as a result of doing so. */
+	div >>= clk_info->div.div;
+	div <<= clk_info->div.div;
 
 	return div;
 }
@@ -401,7 +408,7 @@ static int jz47xx_clk_set_rate(struct clk_hw *hw, unsigned long req_rate,
 		/* update the divide */
 		mask = (1 << clk_info->div.bits) - 1;
 		reg &= ~(mask << clk_info->div.shift);
-		reg |= (div - 1) << clk_info->div.shift;
+		reg |= ((div >> clk_info->div.div) - 1) << clk_info->div.shift;
 
 		/* clear the stop bit */
 		if (clk_info->div.stop_bit != -1)