With the move from 3.0.8 to 3.18 the MMC performance has decreased significantly. This PR addresses this by correctly setting the MSC clocks.

A quick ```dd``` shows the following:

Before:
```
root@ci20:~# dd if=/mnt/sd/test.bin of=/mnt/tmpfs/test.bin bs=4096

32768+0 records in

32768+0 records out

134217728 bytes (134 MB) copied, 12.0821 s, 11.1 MB/s


root@ci20:~# echo 3 > /proc/sys/vm/drop_caches

[  117.210000] bash (606): drop_caches: 3


root@ci20:~# dd if=/mnt/tmpfs/test.bin of=/mnt/sd/test.bin bs=4096

32768+0 records in

32768+0 records out

134217728 bytes (134 MB) copied, 10.9081 s, 12.3 MB/s
```

After:

```
root@ci20:~# dd if=/mnt/sd/test.bin of=/mnt/tmpfs/test.bin bs=4096

32768+0 records in

32768+0 records out
134217728 bytes (134 MB) copied, 6.38894 s, 21.0 MB/s


root@ci20:~# echo 3 > /proc/sys/vm/drop_caches

[  292.870000] bash (574): drop_caches: 3


root@ci20:~# dd if=/mnt/tmpfs/test.bin of=/mnt/sd/test.bin bs=4096

32768+0 records in

32768+0 records out

134217728 bytes (134 MB) copied, 7.10854 s, 18.9 MB/s
```