@aejsmith 2015-07-23
I've added some comments inline on the CGU commit.