@ZubairLK 2015-07-23
Nice catch..

________________________________
From: Alex Smith [notifications@github.com]
Sent: 23 July 2015 12:53
To: MIPS/CI20_linux
Subject: Re: [CI20_linux] Ci20 v3.18 MSC0 Performance Fix (#54)


@paulburton<https://github.com/paulburton> may also want to check this

—
Reply to this email directly or view it on GitHub<https://github.com/MIPS/CI20_linux/pull/54#issuecomment-124071918>.